+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.16.1-dev
 
+* Add a top-level `lastBy()` function that converts an `Iterable` to a `Map` by
+  grouping its elements using a function, keeping the last element for each
+  computed key.
+
 ## 1.16.0
 
 * Add an `Iterable.slices` extension method.